Persistence options
Added persistence options which will mutate the DOM if the clustered table has any rows that have an input or textarea element since these elements will not be persisted by default.   Runs a quick DOM mutation on the currently rendered rows and then updates those indices within the clusterized data. The persistence module will not run if there are no input or textarea elements detected in the markup, but it will check again on a clusterize update to see if any new elements contain an input or textarea.